refactor: removed circular import
