+++ bpy_addon_build/build_context/hooks.py
@@ -1,5 +1,5 @@
 from bpy_addon_build.api import BabContext
-from bpy_addon_build.build_context import BuildContext, console
+from bpy_addon_build.build_context.core import BuildContext, console
 from bpy_addon_build.build_context.hook_definitions import (
     build_action_cleanup,
     build_action_prebuild,
